requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.40
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.2.2
yarl==1.8.1
